Move proxy db to its own file
